break out downloading into its own file. Plus it gets the os and arch info
