+++ Cargo.toml
@@ -28,7 +28,7 @@ rolling-file = "0.2.0"
 secrecy = { version = "0.8.0", features = ["serde"] }
 serde = { version = "1.0.154", features = ["derive"] }
 strum = { version = "0.24.1", features = ["derive"] }
-subtle = "2.5.0"
+subtle = "2.4.1"
 tempfile = "3.4.0"
 serde_yaml = "0.9.19"
 thiserror = "1.0.39"
@@ -44,6 +44,7 @@ yaque = "0.6.4"
 reqwest = "0.11.14"
 tar = "0.4.38"
 flate2 = "1.0.25"
+zip = "0.6.4"
 
 [dev-dependencies]
 insta = { version = "1.28.0", features = ["filters", "json", "yaml"] }
