+++ src/lib.rs
@@ -14,6 +14,7 @@ pub mod api;
 pub mod config;
 pub mod debug;
 pub mod doc;
+pub mod download_fossa_cli;
 pub mod ext;
 pub mod queue;
 pub mod subcommand;
